Update to V1.4
Removed global variables in favor of the setter functions simply return the values. Realized that setters were run to update the variables before using them so why not just have the setter return the values. Seemed redundant. 
Also, added a manual disable function for the front panel button. The way the function worked in the circuit made it look like a fault, but with all zero's. Thus in the ISR, the fault register is read and if all zero, then the circuit was disabled manually, if > 0, then there was indeed a fault.